+++ Source/r_things.c
@@ -63,8 +63,9 @@ static lighttable_t **spritelights;        // killough 1/25/98 made static
 // constant arrays
 //  used for psprite clipping and initializing clipping
 
-short negonearray[MAX_SCREENWIDTH];        // killough 2/8/98:
-short screenheightarray[MAX_SCREENWIDTH];  // change to MAX_*
+// [FG] 32-bit integer math
+int negonearray[MAX_SCREENWIDTH];        // killough 2/8/98:
+int screenheightarray[MAX_SCREENWIDTH];  // change to MAX_*
 
 //
 // INITIALIZATION FUNCTIONS
@@ -303,8 +304,8 @@ vissprite_t *R_NewVisSprite(void)
 //  in posts/runs of opaque pixels.
 //
 
-short   *mfloorclip;
-short   *mceilingclip;
+int   *mfloorclip; // [FG] 32-bit integer math
+int   *mceilingclip; // [FG] 32-bit integer math
 fixed_t spryscale;
 fixed_t sprtopscreen;
 
@@ -819,8 +820,9 @@ void R_SortVisSprites (void)
 void R_DrawSprite (vissprite_t* spr)
 {
   drawseg_t *ds;
-  short   clipbot[MAX_SCREENWIDTH];       // killough 2/8/98:
-  short   cliptop[MAX_SCREENWIDTH];       // change to MAX_*
+  // [FG] 32-bit integer math
+  int   clipbot[MAX_SCREENWIDTH];       // killough 2/8/98:
+  int   cliptop[MAX_SCREENWIDTH];       // change to MAX_*
   int     x;
   int     r1;
   int     r2;
